Correct mistake when filtering subsequent gate searchs

+++ inc/board.h
@@ -43,6 +43,6 @@ void create_block_ind(Board * board, u16 block, u8 ind);
 void break_block(Board * board, u8 x, u8 y);
 void break_block_ind(Board * board, u8 ind);
 u16 seek_block(Board * board, u16 block);
-u16 seek_block_front_msk(Board * board, u16 block, u8 from, u16 msk);
+u8 seek_block_front_msk(Board * board, u16 block, u8 from, u16 msk);
 
 #endif
